server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=0.6.23
pydantic>=2.6.0
python-multipart>=0.0.9
openai>=1.12.0
python-dotenv>=1.0.0
mammoth>=1.8.0
pandas>=2.2.0
sse-starlette>=2.0.0
